refactor: remove extra closing tags
Components without children can be self-closed to avoid the unnecessary extra closing tag. In JSX, closing tags are required when the component has children example `<MyComponent>...</MyComponent>` and if there are no child component between these tags, then this component can be self closed using `<MyComponent />`. It is recommended as it improves readability, and it is more compact to use self-closing for these types of components.